Extract exceptions to their own file
